refactor(medication): split medication-service into prescription/phase/inventory services
medication-service.ts (743 LOC) mixed prescriptions, phases, inventory items,
and inventory transactions. Split into focused services:

- prescription-service.ts: prescription CRUD (add/update/delete/get*)
- phase-service.ts: phase lifecycle and addMedicationToPrescription
- inventory-service.ts: expanded with adjustStock, inventory item/transaction
  CRUD, plus the existing stock-derivation helpers
- medication-builders.ts: shared private record builders

medication-service.ts becomes a barrel re-export so existing callers keep
compiling; a follow-up PR will migrate import sites and delete the barrel.

No behavioural changes; pnpm typecheck, test, lint, and build all green.
